refactor: split partnerresourcesview across sites/brc-analytics and sites/ga2 (#1552)
