--- app/components/content/ga2/index.tsx
@@ -1,1 +1,0 @@
-export { default as SectionAbout } from "./sectionAbout.mdx";
