--- app/components/content/index.tsx
@@ -1,1 +1,0 @@
-export { default as SectionAbout } from "./sectionAbout.mdx";
