+++ sites/brc-analytics/views/PartnerResourcesView/content/sectionAbout.mdx
