+++ sites/ga2/views/PartnerResourcesView/content/sectionAbout.mdx
